+++ microimpute/comparisons/quantile_loss.py
@@ -12,7 +12,7 @@ import numpy as np
 import pandas as pd
 from pydantic import validate_call
 
-from us_imputation_benchmarking.config import QUANTILES, VALIDATE_CONFIG
+from microimpute.config import QUANTILES, VALIDATE_CONFIG
 
 log = logging.getLogger(__name__)
