+++ microimpute/models/matching.py
@@ -8,9 +8,9 @@ import pandas as pd
 from pydantic import validate_call
 from rpy2.robjects import pandas2ri
 
-from us_imputation_benchmarking.config import VALIDATE_CONFIG
-from us_imputation_benchmarking.models.imputer import Imputer, ImputerResults
-from us_imputation_benchmarking.utils.statmatch_hotdeck import (
+from microimpute.config import VALIDATE_CONFIG
+from microimpute.models.imputer import Imputer, ImputerResults
+from microimpute.utils.statmatch_hotdeck import (
     nnd_hotdeck_using_rpy2,
 )
