+++ microimpute/models/ols.py
@@ -9,8 +9,8 @@ import statsmodels.api as sm
 from pydantic import validate_call
 from scipy.stats import norm
 
-from us_imputation_benchmarking.config import VALIDATE_CONFIG
-from us_imputation_benchmarking.models.imputer import Imputer, ImputerResults
+from microimpute.config import VALIDATE_CONFIG
+from microimpute.models.imputer import Imputer, ImputerResults
 
 
 class OLSResults(ImputerResults):
