+++ microimpute/tests/test_models/test_matching.py
@@ -6,10 +6,10 @@ from typing import Dict, List
 import pandas as pd
 from sklearn.datasets import load_iris
 
-from us_imputation_benchmarking.comparisons.data import preprocess_data
-from us_imputation_benchmarking.config import QUANTILES
-from us_imputation_benchmarking.evaluations import *
-from us_imputation_benchmarking.models.matching import Matching
+from microimpute.comparisons.data import preprocess_data
+from microimpute.config import QUANTILES
+from microimpute.evaluations import *
+from microimpute.models.matching import Matching
 
 # Test Method on iris dataset
 iris_data = load_iris()
