+++ microimpute/tests/test_quantile_comparison.py
@@ -12,9 +12,9 @@ from typing import List, Type
 
 import pandas as pd
 
-from us_imputation_benchmarking.comparisons import *
-from us_imputation_benchmarking.config import RANDOM_STATE
-from us_imputation_benchmarking.models import *
+from microimpute.comparisons import *
+from microimpute.config import RANDOM_STATE
+from microimpute.models import *
 
 
 def test_quantile_comparison() -> None:
